fixing the issue of invalid id

+++ server/server/controllers/categoryController/categoryController.js
@@ -1,6 +1,9 @@
 // controllers/CategoryController.js
 const Category = require("../../models/Category");
+const mongoose = require("mongoose");
 const { response } = require("express");
+const ObjectId = mongoose.Types.ObjectId;
+
 class CategoryController {
   // Create a new category
   async createCategory(req, res) {
@@ -24,26 +27,35 @@ class CategoryController {
       const categories = await Category.find({});
       res.status(200).json(categories);
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(400).json({ error: error.message });
     }
   }
 
-  // Get a category by ID
   async getCategoryById(req, res) {
     try {
-      const category = await Category.findById(req.params.id);
+      const id = req.params.id;
+      // Check if the id is a valid ObjectId
+      if (!ObjectId.isValid(id)) {
+        return res.status(400).json({ error: "Invalid ID format" });
+      }
+      const category = await Category.findById(id);
       if (!category) {
         return res.status(404).json({ error: "Category not found" });
       }
       res.status(200).json(category);
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(400).json({ error: error.message });
     }
   }
 
   // Update a category by ID
   async updateCategory(req, res) {
     try {
+      const id = req.params.id;
+      // Check if the id is a valid ObjectId
+      if (!ObjectId.isValid(id)) {
+        return res.status(400).json({ error: "Invalid ID format" });
+      }
       const name = req?.body?.name;
       const cate = await Category.findOne({ name });
       if (cate) {
@@ -66,13 +78,22 @@ class CategoryController {
   // Delete a category by ID
   async deleteCategory(req, res) {
     try {
+      const id = req.params.id;
+      // Check if the id is a valid ObjectId
+      if (!ObjectId.isValid(id)) {
+        return res.status(400).json({ error: "Invalid ID format" });
+      }
+      const categories = await Category.findById(id);
+      if (!categories) {
+        return res.status(404).json({ error: "Category not found" });
+      }
       const category = await Category.findByIdAndDelete(req.params.id);
       if (!category) {
         return res.status(404).json({ error: "Category not found" });
       }
       res.status(200).json({ message: "Category deleted" });
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(400).json({ error: error.message });
     }
   }
 }
